starfield cuda: fix vertex shader so that particle size is correct(*)
I did things like test that as I move the camera around, getting
closer and farther away, the size of a rendered particle stays
constant relative to other objects in the scene.

(*) Note that without actually rendering a real sphere with many real
points it'll be difficult to get the geometry exactly perfect when the
image of the sphere fills most of the camera and especially when it's
off-center.