analysislib.combine: fix added dimension bug
if there are any columns common in both dataframes the result
seems to be that the pd.concat resizes the contained values
by adding an extra dimenstion.
df['x'].values.ndim = 1 becomes = 2 (for version of
pandas < 0.14). To work around this, remove any columns
in the csv dataframe that exists in the trajectory
dataframe