sid-analysis: move more analysis to matlab
conceptually I create an iddata object which holds
the input/output data for each trial (and for the means).
This allows future processing to use the
MATLAB sid-toolbox merge operation to either merge the
trial data or merge the models.

Doing the bode and polezero plots in matlab also allows
showing confidence intervals on the plots.