sid-analysis: individual fly models per trajectory
refactor code so there is more flexibility in the choice
of modelling and validation data. Use this to run the estimation
procedure on each individual fly (using pre and perturb data) and
not just on the mean stimulus/response.

the current operation is thus
 * collect flies which completed the perturbation
 * calculate the mean response over the perturb period
 * try to identify many different model order (and methods),
   tf44, oe441, etc
 * for those models with a good fit over the mean data
   * run that same model order/method estimation procedure
     on the individual flies
   * plot individual fly models

save iddata per fly, the merged iddata per condition, and
the mean iddata per condition