analysislib: add get_condtion_object and fallback
the fallback uses some regexs to heuristically decide
if an old /a/b/c condition string is a perturbation,
rotation or conflict experiment.

this makes it easy for certain analysis scripts
to easily test if a condition is a certain type, e.g.

obj = combine.get_condition_object(current_condition)
if obj.is_type('rotation'):
    #do some only rotation plots